Fix compile: remove BuildConfig usage; compute version via PackageManager; cast slider value to Int

+++ app/src/main/java/com/n0tez/app/NoteEditorActivity.kt
@@ -36,7 +36,7 @@ class NoteEditorActivity : AppCompatActivity() {
             sliderTransparency.addOnChangeListener { _, value, _ ->
                 val alpha = value / 100f
                 noteContainer.alpha = alpha
-                saveTransparencyPreference(value)
+                saveTransparencyPreference(value.toInt())
             }
             
             // Setup text editor
